+++ astrbot/cli/REAMDE.MD
@@ -1,0 +1,23 @@
+#Astrbot.cli包
+
+组织结构：
+cli
+├── __init__.py
+├── init.py 附属于cli根命令组，负责初始化astrbot 根目录结构
+└── run.py 附属于cli根命令组 ，负责启动astrbot
+|————conf 子命令组
+    ├── __init__.py 附属于cli.conf命令组，负责导出子命令组conf
+    ├── __main__.py 附属于cli.conf命令组，命令组入口，需采用绝对导入
+
+1. 命令组对应一个文件夹
+比如cli文件夹对应cli根命令组
+conf文件夹对应conf 子命令组
+
+2. __main__.py只能放命令组和基本内置命令，比如自定义的help命令之类的
+且不能使用相对导入，设计为了可以直接使用python -m 运行，和直接运行。
+其余设计为被导入使用的统一采用相对导入，且延迟导入
+
+3. __init__.py 仅导出命令组
+
+4. 命令.py：以命令名命名
+
