+++ astrbot/cli/__init__.py
@@ -1,1 +1,6 @@
 __version__ = "3.5.8"
+from .__main__ import cli
+
+__all__ = [
+    "cli"
+]
