--- astrbot/cli/commands/__init__.py
@@ -1,6 +1,0 @@
-from .cmd_init import init
-from .cmd_run import run
-from .cmd_plug import plug
-from .cmd_conf import conf
-
-__all__ = ["init", "run", "plug", "conf"]
