+++ astrbot/cli/conf/__init__.py
@@ -1,0 +1,3 @@
+from .__main__ import conf
+
+__all__ = ["conf"]
