+++ astrbot/cli/plug/__init__.py
@@ -1,0 +1,3 @@
+from .__main__ import plug
+
+__all__ = ["plug"]
