+++ astrbot/cli/plug/__main__.py
@@ -1,0 +1,24 @@
+import click
+
+@click.group()
+def plug():
+    """插件管理"""
+    pass
+
+from astrbot.cli.plug.new import new
+from astrbot.cli.plug.search import search
+from astrbot.cli.plug.list import list_plugins
+from astrbot.cli.plug.install import install
+from astrbot.cli.plug.update import update
+from astrbot.cli.plug.remove import remove
+
+plug.add_command(new)
+plug.add_command(search)
+plug.add_command(list_plugins)
+plug.add_command(install)
+plug.add_command(update)
+plug.add_command(remove)
+
+
+if __name__ == "__main__":
+    plug()
