+++ astrbot/cli/utils/__init__.py
@@ -5,6 +5,15 @@ from .basic import (
 )
 from .plugin import get_git_repo, manage_plugin, build_plug_list, PluginStatus
 from .version_comparator import VersionComparator
+from .valutils import (
+    validate_log_level,
+    validate_dashboard_port,
+    validate_dashboard_username,
+    validate_dashboard_password,
+    validate_callback_api_base,
+    validate_timezone,
+)
+
 
 __all__ = [
     "get_astrbot_root",
@@ -15,4 +24,10 @@ __all__ = [
     "build_plug_list",
     "VersionComparator",
     "PluginStatus",
+    "validate_log_level",
+    "validate_dashboard_port",
+    "validate_dashboard_username",
+    "validate_dashboard_password",
+    "validate_callback_api_base",
+    "validate_timezone",
 ]
